challenge/challenge/src/main/java/com/dws/challenge/controller/TransferController.java
package com.dws.challenge.model;

import com.fasterxml.jackson.annotation.JsonInclude;
import com.fasterxml.jackson.annotation.JsonPropertyOrder;
import com.fasterxml.jackson.databind.annotation.JsonSerialize;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@AllArgsConstructor
@NoArgsConstructor
@Data
@Builder
@JsonSerialize
@JsonInclude(JsonInclude.Include.NON_NULL)
@JsonPropertyOrder({
	"accountFromId",
	"accountToId"
	"amount"})
public class TransferRequest {
	private Long accountFromId;
	private Long accountToId;
	private double amount;
}